self review and code documentation
