+++ integration_test/ESI/cosim/lowlevel_cosim_only.sv.py
@@ -1,3 +1,5 @@
+# Test driver for mmio cosim.
+
 import esi_cosim
 import capnp
 import sys
