+++ lib/Dialect/ESI/cosim/Cosim_DpiPkg.sv
@@ -76,26 +76,29 @@ import "DPI-C" sv2cCosimserverEpTryGet =
 
 // --------------------- Low Level interface -----------------------------------
 
+/// Register an MMIO module. Just checks that there is only one instantiated.
 import "DPI-C" sv2cCosimserverMMIORegister =
   function int cosim_mmio_register();
 
+/// Read MMIO function pair. Assumes that reads come back in the order in which
+/// they were requested.
 import "DPI-C" sv2cCosimserverMMIOReadTryGet =
   function int cosim_mmio_read_tryget(
     output int address
   );
-
 import "DPI-C" sv2cCosimserverMMIOReadRespond =
   function void cosim_mmio_read_respond(
     input longint data,
     input byte    error
   );
 
+/// Write MMIO function pair. Assumes that write errors come back in the order
+/// in which the writes were issued.
 import "DPI-C" sv2cCosimserverMMIOWriteTryGet =
   function int cosim_mmio_write_tryget(
     output int     address,
     output longint data
   );
-
 import "DPI-C" sv2cCosimserverMMIOWriteRespond =
   function void cosim_mmio_write_respond(
     input byte    error
