+++ lib/Dialect/ESI/cosim/cosim_dpi_server/CMakeLists.txt
@@ -16,8 +16,7 @@ if(ESI_COSIM)
     DpiEntryPoints.cpp
     Server.cpp
     Endpoint.cpp
-    LowLevel.cpp)
-
+  )
   set_target_properties(EsiCosimDpiServer
       PROPERTIES
           LIBRARY_OUTPUT_DIRECTORY ${CMAKE_BINARY_DIR}/lib
