--- lib/Dialect/ESI/cosim/cosim_dpi_server/LowLevel.cpp
@@ -1,14 +1,0 @@
-//===- LowLevel.cpp - Cosim low level RPC - simulation bridge ---*- C++ -*-===//
-//
-// Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
-// See https://llvm.org/LICENSE.txt for license information.
-// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
-//
-//===----------------------------------------------------------------------===//
-
-#include "circt/Dialect/ESI/cosim/LowLevel.h"
-
-using namespace circt::esi::cosim;
-
-LowLevel::LowLevel() {}
-LowLevel::~LowLevel() {}
